+++ pytsite/file_storage_odm/__init__.py
@@ -1,4 +1,4 @@
-"""PytSite File ODM Storage.
+"""PytSite ODM File Storage.
 """
 from . import _model as model, _field as field
 from ._driver import Driver
@@ -13,7 +13,7 @@ def _init():
     from . import _model, _eh
 
     # Register ODM models
-    odm.register_model('file_other', _model.AnyFileODMEntity)
+    odm.register_model('file', _model.AnyFileODMEntity)
     odm.register_model('file_image', _model.ImageFileODMEntity)
 
     # Event handlers
@@ -23,5 +23,6 @@ def _init():
         '/image/resize/<int:width>/<int:height>/<string(length=2):p1>/<string(length=2):p2>/<string:filename>',
         'pytsite.file_storage_odm@image',
     )
+
 
 _init()